v1.5.0-beta.1: Multi-channel pipeline, pre-capture validation, and 16 bug fixes
Major features:
- RGB color camera support through entire pipeline (capture, averaging,
  ΔR/R, display, analysis, export) with Bayer demosaic for pypylon
- Pre-capture validation system with read-only preflight checks
  (exposure, stability, focus, hardware, TEC) and dismissable dialog
- FPS optimizer: 3-step auto-optimization (LED duty → FPS → exposure)
- FFC UI button for Boson/FLIR thermal cameras
- RGB thermoreflectance per-channel analysis module
- Float64 pipeline averaging for >15 significant digits of precision

Structural changes:
- CameraFrame generalized with channels/bit_depth; CameraInfo with pixel_format
- Pipeline processing hooks (pre_capture_hooks, post_average_hooks)
- Session schema v3 with multi-channel metadata and preflight traceability
- Shared image_metrics module for MetricsService and PreflightValidator

Bug fixes (16):
- Thread safety: AF and optimizer UI updates marshaled to main thread
- Stale AF hook closure now resolves hardware at execution time
- Drift correction handles RGB frames (luminance for shift estimation)
- 8-bit→12-bit scaling corrected from <<4 to *4095//255
- Transient/movie pipelines: accumulator shapes, dark masks, and
  apply_shift all handle multi-channel data correctly
- Export preserves float64 for HDF5; TIFF axes correct for 3-channel
- Analysis and overlay reduce RGB to luminance for morphology ops

+++ acquisition/live.py
@@ -214,7 +214,7 @@ class LiveProcessor:
 
             # ---- ΔR/R ----
             cold_safe = np.where(cold_ema > 1.0, cold_ema, 1.0)
-            drr = ((hot_ema - cold_ema) / cold_safe).astype(np.float32)
+            drr = ((hot_ema - cold_ema) / cold_safe).astype(np.float64)
 
             # ---- SNR estimate ----
             snr = self._estimate_snr(drr)
@@ -283,7 +283,7 @@ class LiveProcessor:
             time.sleep(cfg.trigger_delay_ms / 1000.0)
 
     def _grab_avg(self, n: int, cfg: LiveConfig) -> Optional[np.ndarray]:
-        """Grab N frames and return their float32 average, with optional ROI crop.
+        """Grab N frames and return their float64 average, with optional ROI crop.
 
         If the frame shape changes mid-accumulation (e.g. because set_resolution()
         was called on a simulated camera), the accumulator is restarted so we never
@@ -316,7 +316,7 @@ class LiveProcessor:
 
         if acc is None or count == 0:
             return None
-        return (acc / count).astype(np.float32)
+        return (acc / count).astype(np.float64)
 
     def _synthetic_frame(self, cfg: LiveConfig) -> np.ndarray:
         """Generate a synthetic camera frame for simulation."""
